third_party/WebKit/LayoutTests/animations/custom-properties/length-percentage-type-mismatch-no-interpolation-expected.txt

CONSOLE WARNING: line 24: Invalid keyframe value for property --length: 10%
CONSOLE WARNING: line 35: Invalid keyframe value for property --percentage: 10px
This is a testharness.js-based test.
PASS <length> properties don't accept percentages in animation keyframes
PASS <percentage> properties don't accept lengths in animation keyframes
Harness: the test ran to completion.
